Move MergePolicyForLevel type to the module where it is used
It is not used in MergingRun. And it should not live there because
MergingRun is shared between level merges and union merges, but a policy
for a level is obviously level specific.
